requirements.txt
Flask-Cors~=3.0.10
Flask-Migrate~=3.1.0
Flask-Script~=2.0.6
Flask-SQLAlchemy~=2.5.1
Flask~=2.0.2
PyJWT~=2.4.0
PyMySQL~=1.0.2
SQLAlchemy~=1.4.26
alembic~=1.7.4
marshmallow~=3.14.0
google_auth~=2.6.6
SAValidation~=0.4.1
requests~=2.27.1
pusher~=2.1.4
boto3~=1.21.22
diskcache~=5.4.0
python-memcached~=1.59